8312979: Fix assembler_aarch64.hpp after JDK-8311847

+++ src/hotspot/cpu/aarch64/assembler_aarch64.hpp
@@ -653,7 +653,7 @@ class Address {
   static bool offset_ok_for_sve_immed(int64_t offset, int shift, int vl /* sve vector length */) {
     if (offset % vl == 0) {
       // Convert address offset into sve imm offset (MUL VL).
-      int64_t sve_offset = offset / vl;
+      int32_t sve_offset = (int32_t)(offset / vl);
       if (((-(1 << (shift - 1))) <= sve_offset) && (sve_offset < (1 << (shift - 1)))) {
         // sve_offset can be encoded
         return true;
